requirements.txt
streamlit
groq
pypdf2
python-dotenv
sentence-transformers
faiss-cpu
numpy
matplotlib
pandas
